healthcare-imaging-ai/requirements.txt
torch
torchvision
opencv-python
numpy
Pillow
matplotlib
